requirements.txt
aiogram==2.12
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
emoji==1.2.0
gino==1.0.1
idna==3.1
multidict==5.1.0
Pillow==8.1.2
python-dotenv==0.15.0
pytz==2021.1
pyzbar==0.1.8
SQLAlchemy==1.3.23
typing-extensions==3.7.4.3
yarl==1.6.3
